refactor: split gen-llms into markdown and guide content modules
Extract the minimal markdown renderer and guide prose constants from the
monolithic gen-llms module so prose edits no longer collide with catalog
builders and llms surface assembly. Keep $scripts/gen-llms as the public
re-export entrypoint.

Route the new sibling modules through the docs lane and component_journeys
content-hash inputs, and assert the runtime export surface plus docs-lane
scheduling so path routing cannot false-green after a prose-only edit.
